coordinates are doubles actually

+++ entities/Texture_t.h
@@ -13,8 +13,8 @@ class Texture_t{
         unsigned int size_y_;
         Vec3f** img_;
 
-        Vec3f get(unsigned int (&xy)[2]) const {};
-        Vec3f get_nn(unsigned int (&xy)[2]) const {};
+        Vec3f get(double (&xy)[2]) const {};
+        Vec3f get_nn(double (&xy)[2]) const {};
 };
 
 #endif
